Move socket metrics to single file (#618)
