+++ ports/common/include/iot_network_openssl.h
@@ -165,9 +165,9 @@ IotNetworkError_t IotNetworkOpenssl_Close( void * pConnection );
 IotNetworkError_t IotNetworkOpenssl_Destroy( void * pConnection );
 
 /**
- * @brief Used by metrics to retrieve remote server and port of a connection.
+ * @brief Used by metrics to retrieve the socket (file descriptor) associated with
+ * a connection.
  */
-void IotNetworkOpenssl_GetServerInfo( void * pConnection,
-                                      IotMetricsTcpConnection_t * pServerInfo );
+int IotNetworkOpenssl_GetSocket( void * pConnection );
 
 #endif /* ifndef IOT_NETWORK_OPENSSL_H_ */
