+++ ports/common/src/iot_network_mbedtls.c
@@ -32,12 +32,6 @@
 #include <stdio.h>
 #include <string.h>
 
-#if !defined( _WIN32 ) && !defined( _WIN64 )
-    #include <arpa/inet.h>
-    #include <netdb.h>
-    #include <sys/socket.h>
-#endif
-
 /* mbed TLS network include. */
 #include "iot_network_mbedtls.h"
 
@@ -1198,100 +1192,12 @@ IotNetworkError_t IotNetworkMbedtls_Destroy( void * pConnection )
 
 /*-----------------------------------------------------------*/
 
-void IotNetworkMbedtls_GetServerInfo( void * pConnection,
-                                      IotMetricsTcpConnection_t * pServerInfo )
+int IotNetworkMbedtls_GetSocket( void * pConnection )
 {
-    int status = 0, portLength = 0;
-    struct sockaddr_storage server = { 0 };
-    socklen_t length = sizeof( struct sockaddr_storage );
-    const void * pServerAddress = NULL;
-    char * pAddressStart = NULL;
-    const char * pPortFormat = NULL;
-    uint16_t remotePort = 0;
-    size_t addressLength = 0;
-    const _networkConnection_t * pNetworkConnection = NULL;
-
     /* Cast function parameter to correct type. */
-    if( ( pConnection != NULL ) && ( pServerInfo != NULL ) )
-    {
-        pNetworkConnection = pConnection;
+    _networkConnection_t * const pNetworkConnection = pConnection;
 
-        /* Get peer info. */
-        status = getpeername( pNetworkConnection->networkContext.fd,
-                              ( struct sockaddr * ) &server,
-                              &length );
-    }
-    else
-    {
-        status = IOT_NETWORK_BAD_PARAMETER;
-    }
-
-    if( status == 0 )
-    {
-        /* Calculate the pointer to the IP address and get the remote port based
-         * on protocol version. */
-        if( server.ss_family == AF_INET )
-        {
-            /* IPv4. */
-            pServerAddress = &( ( ( struct sockaddr_in * ) &server )->sin_addr );
-            remotePort = ntohs( ( ( struct sockaddr_in * ) &server )->sin_port );
-
-            /* Print the IPv4 address at the start of the address buffer. */
-            pAddressStart = pServerInfo->pRemoteAddress;
-            addressLength = IOT_METRICS_IP_ADDRESS_LENGTH;
-            pPortFormat = ":%hu";
-        }
-        else
-        {
-            /* IPv6. */
-            pServerAddress = &( ( ( struct sockaddr_in6 * ) &server )->sin6_addr );
-            remotePort = ntohs( ( ( struct sockaddr_in6 * ) &server )->sin6_port );
-
-            /* Enclose the IPv6 address with []. */
-            pServerInfo->pRemoteAddress[ 0 ] = '[';
-            pAddressStart = pServerInfo->pRemoteAddress + 1;
-            addressLength = IOT_METRICS_IP_ADDRESS_LENGTH - 1;
-            pPortFormat = "]:%hu";
-        }
-
-        /* Convert IP address to text. */
-        if( inet_ntop( server.ss_family,
-                       pServerAddress,
-                       pAddressStart,
-                       addressLength ) != NULL )
-        {
-            /* Add the port to the end of the address. */
-            addressLength = strlen( pServerInfo->pRemoteAddress );
-
-            portLength = snprintf( &( pServerInfo->pRemoteAddress[ addressLength ] ),
-                                   7,
-                                   pPortFormat,
-                                   remotePort );
-
-            if( portLength > 0 )
-            {
-                pServerInfo->addressLength = addressLength + ( size_t ) portLength;
-
-                IotLogInfo( "(Socket %d) Collecting network metrics for %s.",
-                            pNetworkConnection->networkContext.fd,
-                            pServerInfo->pRemoteAddress );
-            }
-            else
-            {
-                IotLogError( "(Socket %d) Failed to add port to IP address buffer." );
-            }
-        }
-        else
-        {
-            IotLogError( "(Socket %d) Failed to convert IP address to text format.",
-                         pNetworkConnection->networkContext.fd );
-        }
-    }
-    else
-    {
-        IotLogError( "(Socket %d) Failed to query peer name.",
-                     pNetworkConnection->networkContext.fd );
-    }
+    return pNetworkConnection->networkContext.fd;
 }
 
 /*-----------------------------------------------------------*/
